+++ Kokoro/API/CardRenderHook.cs
@@ -11,6 +11,7 @@ public partial interface IKokoroApi
 
 public interface ICardRenderHook
 {
+	bool ShouldDisableCardRenderingTransformations(G g, Card card) => false;
 	Vec ModifyTextCardScale(G g, Card card) => Vec.One;
 	Matrix ModifyNonTextCardRenderMatrix(G g, Card card, List<CardAction> actions) => Matrix.Identity;
 	Matrix ModifyCardActionRenderMatrix(G g, Card card, List<CardAction> actions, CardAction action, int actionWidth) => Matrix.Identity;